backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
anthropic
openai
httpx
python-dotenv
aiofiles
bcrypt
